pyproject.toml
[tool.poetry]
name = "hotel"
version = "0.1.0"
description = ""
authors = ["Shakhboz Usmanov <shakhbozusmanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.4.0"
flower = "^2.0.1"
pillow = "^10.3.0"
sqladmin = "^0.17.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
black = "^24.4.2"
flake8 = "^7.1.0"
autoflake = "^2.3.1"
isort = "^5.13.2"
pyright = "^1.1.369"
python-json-logger = "^2.0.7"
bcrypt = "^4.1.3"
sentry-sdk = {extras = ["fastapi"], version = "^2.7.1"}
fastapi-versioning = "^0.10.0"
gunicorn = "^22.0.0"
poetry-plugin-export = "^1.8.0"
pandas = "^2.2.2"
prometheus-fastapi-instrumentator = "^7.0.0"
asgiref = "^3.8.1"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
